requirements.txt
PyYaml>=5.4.1
tabulate